requirements.txt
fastapi==0.121.3
uvicorn[standard]==0.38.0
pydantic==2.12.4
python-dotenv>=1.0,<2.0

langchain==1.0.8
langchain-core==1.1.0
langchain-community==0.4.1
langchain-openai==1.0.3
langgraph==1.0.3

langchain-chroma==1.0.0
chromadb==1.3.5

tiktoken==0.12.0
pypdf==6.3.0
python-docx==1.2.0
rank-bm25==0.2.2

python-multipart==0.0.20
aiofiles==24.1.0
streamlit==1.40.2

pymysql>=1.1.1
redis==7.0.0
